This commit was generated by cvs2svn to compensate for changes in r15375, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@15376 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/traps.c
@@ -238,6 +238,6 @@ DWORD traps_handler(void)
         p = p->next;
     }
 
-    return -1;
+    return (DWORD)-1;
 }
